+++ scripts/db/watch_table.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+export PGPASSWORD='mysecretpassword'
+export POSTGRESQL_URL=postgres://postgres:mysecretpassword@localhost:5432/devdb?sslmode=disable
+
+watch "psql -h localhost -U postgres -d devdb -p 5432 -c 'SELECT * FROM users LIMIT 10;'"
